% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_reference.R
\name{LoadReference}
\alias{LoadReference}
\title{LoadReference}
\usage{
LoadReference(reference.gene.xls)
}
\arguments{
\item{reference.gene.xls}{File containing the real Gene Symbol (colname 'gene_symbol')}
}
\value{
ref_data a data frame consists of two columns ["fake_gene_symbol", "gene_symbol"].
"fake_gene_symbol" are the genes that we match in the gene list, "gene_symbol" are the real names that we match
}
\description{
LoadReference
}
\examples{
ref_data <- LoadReference(reference.gene.xls)
}
