% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cds.R
\name{RenameGenesCellDataSet}
\alias{RenameGenesCellDataSet}
\title{RenameGenesCellDataSet}
\usage{
RenameGenesCellDataSet(cds_data, ref_data)
}
\arguments{
\item{cds_data}{CellDataSet object}

\item{ref_data}{a data frame consists of two columns ["fake_gene_symbol", "gene_symbol"].
"fake_gene_symbol" are the genes that we match in the gene list, "gene_symbol" are the real names that we match}
}
\value{
CellDataSet object
}
\description{
rename gene symbol of CellDataSet object
}
\examples{
new_cds_data <- RenameGenesCellDataSet(cds_data, ref_data)
}
