% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_matrix.R
\name{RenameGenesMatrix}
\alias{RenameGenesMatrix}
\title{RenameGenesMatrix}
\usage{
RenameGenesMatrix(matrix, ref_data, ncol, if_rownames = FALSE)
}
\arguments{
\item{matrix}{A matrix containing the genetic symbol}

\item{ref_data}{a data frame consists of two columns ["fake_gene_symbol", "gene_symbol"].
"fake_gene_symbol" are the genes that we match in the gene list, "gene_symbol" are the real names that we match}

\item{ncol}{The name of the gene column in the matrix}

\item{if_rownames}{Whether the matrix row name needs to be replaced (TRUE/FALSE)}
}
\value{
new matrix
}
\description{
Rename gene symbols for gene matrix
}
\examples{
new_matrix <- RenameGenesMatrix(matrix, ref_data, "gene", if_rownames=FALSE)
}
