% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_list.R
\name{RenameReference}
\alias{RenameReference}
\title{RenameReference}
\usage{
RenameReference(ref_data, gene_list)
}
\arguments{
\item{ref_data}{a data frame consists of two columns ["fake_gene_symbol", "gene_symbol"].
"fake_gene_symbol" are the genes that we match in the gene list, "gene_symbol" are the real names that we match}

\item{gene_list}{a gene symbol list}
}
\value{
new gene list
}
\description{
rename gene list accroding to a ref data
}
\examples{
new_gene_list <- RenameReference(ref_data, gene_list)
}
