% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_tree.R
\name{RenameTree}
\alias{RenameTree}
\title{RenameTree}
\usage{
RenameTree(treefile, newfile, ref_data)
}
\arguments{
\item{treefile}{newick format trees. Some genes are 'si-ch73-1a9.3', but they're actually 'si:ch73-1a9.3'}

\item{newfile}{newick format trees. Gene symbol with ":" will be replace by '"si:ch73-1a9.3"'}

\item{ref_data}{a data frame consists of two columns ["fake_gene_symbol", "gene_symbol"].
"fake_gene_symbol" are the genes that we match in the gene list, "gene_symbol" are the real names that we match}
}
\value{
new s4 object
}
\description{
RenameTree
}
\examples{
RenameTree(treefile, newfile, ref_data)
}
