% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_s4.R
\name{RenameGenesSeurat}
\alias{RenameGenesSeurat}
\title{Rename for Seurat S4 object}
\usage{
RenameGenesSeurat(obj = ls.Seurat[[i]], assay = "RNA")
}
\arguments{
\item{assay}{}
}
\value{

}
\description{
Rename for Seurat S4 object
}
